feat(client): separate cookie logic
