+++ .gitignore
@@ -14,6 +14,8 @@ repo
 .venv
 *.pyc
 */__pycache__/*
+**/.ruff_cache/
+**/test/units/*.json
 .coverage
 .env%
