+++ interface/src/hooks/useGoogleAuth.ts
@@ -1,27 +1,4 @@
-import { call_create_author_agent } from '@api/sc/agents/googleAuthAgent';
 import { useState, useCallback, useEffect } from 'react';
-
-// generate random session
-export const generateSessionId = (): string => {
-  return 'session_' + Math.random().toString(36).substr(2, 9) + '_' + Date.now();
-};
-
-export const setCookie = (name: string, value: string, days: number = 1) => {
-  const expires = new Date();
-  expires.setTime(expires.getTime() + days * 24 * 60 * 60 * 1000);
-  document.cookie = `${name}=${value};expires=${expires.toUTCString()};path=/;SameSite=Lax`;
-};
-
-export const getCookie = (name: string): string | null => {
-  const nameEQ = name + "=";
-  const ca = document.cookie.split(';');
-  for (let i = 0; i < ca.length; i++) {
-    let c = ca[i];
-    while (c.charAt(0) === ' ') c = c.substring(1, c.length);
-    if (c.indexOf(nameEQ) === 0) return c.substring(nameEQ.length, c.length);
-  }
-  return null;
-};
 
 
 export const useGoogleAuth = () => {
@@ -32,7 +9,7 @@ export const useGoogleAuth = () => {
     setIsLoading(true);
     setError(null);
     // use oauth2.0 technology
-    // send request to google(get code param)
+    // send request to google(get code query param)
     const clientId = process.env.GOOGLE_CLIENT_ID;
     const redirectUri = 'http://localhost:3033/auth/google/callback';
     const scopes = [
