+++ interface/src/pages/Auth/Callback/Google/google.tsx
@@ -2,7 +2,7 @@
 import { useEffect, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 import { call_create_author_agent } from '@api/sc/agents/googleAuthAgent';
-import { generateSessionId, setCookie } from '@hooks/useGoogleAuth';
+import { generateSessionId, setCookie } from '@utils';
 
 
 export const GoogleCallback = () => {
