+++ interface/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './refSetter';
 export * from './throttle';
+export * from './cookie';
